+++ src/components/FilterNavigation/index.js
@@ -1,0 +1,1 @@
+export { default as FilterNavigation } from './FilterNavigation';