+++ src/components/SearchTextareaField/SearchTextareaField.js
@@ -17,6 +17,7 @@ const propTypes = {
   onChange: PropTypes.func,
   onChangeIndex: PropTypes.func,
   onSubmitSearch: PropTypes.func.isRequired,
+  placeholder: PropTypes.string,
   searchableIndexes: PropTypes.arrayOf(PropTypes.shape({
     label: PropTypes.string,
     value: PropTypes.string,
@@ -32,6 +33,7 @@ const SearchTextareaField = ({
   onChange,
   loading,
   searchableIndexes,
+  placeholder,
   onChangeIndex,
   selectedIndex,
   disabled,
@@ -65,6 +67,7 @@ const SearchTextareaField = ({
         onChange={onChangeIndex}
         selectClass={css.select}
         value={selectedIndex}
+        placeholder={placeholder}
       />
       <TextArea
         {...rest}