+++ src/constants/index.js
@@ -4,3 +4,4 @@ export * from './searchableIndexesMap';
 export * from './searchResultsListColumns';
 export * from './filterConfig';
 export * from './sortOrders';
+export * from './navigationSegments';