+++ src/constants/rawSearchableIndexes.js
@@ -1,6 +1,6 @@
 import { searchableIndexesValues } from './searchableIndexesValues';
 
-export const rawSearchableIndexes = [
+export const rawDefaultSearchableIndexes = [
   { label: 'ui-marc-authorities.keyword', value: '' },
   { label: 'ui-marc-authorities.identifier', value: searchableIndexesValues.IDENTIFIER },
   { label: 'ui-marc-authorities.personalName', value: searchableIndexesValues.PERSONAL_NAME },
@@ -10,5 +10,15 @@ export const rawSearchableIndexes = [
   { label: 'ui-marc-authorities.uniformTitle', value: searchableIndexesValues.UNIFORM_TITLE },
   { label: 'ui-marc-authorities.subject', value: searchableIndexesValues.SUBJECT },
   { label: 'ui-marc-authorities.childrenSubjectHeading', value: searchableIndexesValues.CHILDREN_SUBJECT_HEADING },
+  { label: 'ui-marc-authorities.genre', value: searchableIndexesValues.GENRE },
+];
+
+export const rawBrowseSearchableIndexes = [
+  { label: 'ui-marc-authorities.personalName', value: searchableIndexesValues.PERSONAL_NAME },
+  { label: 'ui-marc-authorities.corporateConferenceName', value: searchableIndexesValues.CORPORATE_CONFERENCE_NAME },
+  { label: 'ui-marc-authorities.geographicName', value: searchableIndexesValues.GEOGRAPHIC_NAME },
+  { label: 'ui-marc-authorities.nameTitle', value: searchableIndexesValues.NAME_TITLE },
+  { label: 'ui-marc-authorities.uniformTitle', value: searchableIndexesValues.UNIFORM_TITLE },
+  { label: 'ui-marc-authorities.subject', value: searchableIndexesValues.SUBJECT },
   { label: 'ui-marc-authorities.genre', value: searchableIndexesValues.GENRE },
 ];